requirements.txt
Django==1.10.6
django-tinymce==2.7.0
Markdown==2.6.10
olefile==0.44
Pillow==4.3.0
